Arrumado alguns problema de interface (acentuacoes) e sistema de persistencia de login

+++ MackTIA/Entities/Absence.swift
@@ -16,7 +16,7 @@ struct Absence {
     var dadas:Int
     var disciplina:String
     var faltas:Int
-    var percentual:Int
+    var percentual:Float
     var permit:Int
     var permit20:Int
     var turma:String
